api/requirements.txt
flask
psycopg2-binary
flask_cors
